Implement the select race transaction
Implement the select race transaction so we can include races in
characters.

This has a lot of code in it related to ensuring everything matches the
given type and exact names presented so requests and responses can have
a very precise expectation of them.  We have 3 components so far:

- Transactions;
- Entities;
- Formatters.

Normally, we don't need formatters, but since we want to keep the
entities clear of presentation logic and we need to return an actual
object back to the handler (likely a presenter object), I decided to
format the entity objects in a very specific way to maintain this
requirement.

I may end up changing when to use factories in the transactions.  It
almost doesn't make sense to pass factories to the transactions, only
formatters and handlers.  I may also reduce the number of parameters to
one and have them detect a formatter factory, handler factory and,
eventually, a gateway factory to reduce parameters.  Not sure yet.

I also may create subcomponents for the errors to specify which ones
they go to.  I may give each component an "errors" subcomponent.  I
haven't decided which route would work better.

In any case, here's hoping this continues to go as well as it has so
far.

Signed-off-by: Ryan Y <ryayak1460@protingumas.org>