Implement the rest of the subraces
Implement the rest of the subraces for the select subrace transaction
so we can select wood elfs, drow, lightfeet, stouts, forest gnomes and
rock gnomes for our characters if we so please.

Next up, I have to start writing the errors and exceptional cases for
this transaction.  Not many I can think of, but they still exist.

Signed-off-by: Ryan Y <ryayak1460@protingumas.org>